requirements.txt
streamlit>=1.22.0
psycopg2-binary>=2.9.6
openpyxl>=3.1.2